* dev:
  Fix 16-bit/32-bit compatability check
  Tweaks to incompatible VM checks and messages
  Handle mixed 32-bit/16-bit configuration (e.g. 32-bit IDE and 16-bit VM).
  Edit SERIAL_PROTOCOL doc

+++ vm/runtime.c
@@ -1105,6 +1105,7 @@ static void processShortMessage() {
 		sendPingNow(chunkIndex); // send a ping to acknowledge
 		break;
 	case startAllMsg:
+		if (0 != chunkIndex) break; // ignore msg from later VM
 		startAll();
 		break;
 	case stopAllMsg:
@@ -1119,6 +1120,7 @@ static void processShortMessage() {
 		sendVarNames();
 		break;
 	case clearVarsMsg:
+		if (0 != chunkIndex) break; // ignore msg from later VM
 		clearAllVariables();
 		memClear();
 		break;
@@ -1126,6 +1128,7 @@ static void processShortMessage() {
 		sendChunkCRC(chunkIndex);
 		break;
 	case getAllCRCsMsg:
+		if (0 != chunkIndex) break; // ignore msg from later VM
 		sendPingNow(chunkIndex); // send a ping to acknowledge receipt
 		sendAllCRCs();
 		break;
@@ -1133,10 +1136,12 @@ static void processShortMessage() {
 		sendVersionString();
 		break;
 	case getAllCodeMsg:
+		if (0 != chunkIndex) break; // ignore msg from later VM
 		sendPingNow(chunkIndex); // send a ping to acknowledge receipt
 		sendAllCode();
 		break;
 	case deleteAllCodeMsg:
+		if (0 != chunkIndex) break; // ignore msg from later VM
 		deleteAllChunks();
 		memClear();
 		primMBDisplayOff(0, NULL);
